+++ src/BiotSavart/longrange/finufft.jl
@@ -79,6 +79,17 @@ expected_period(::FINUFFTBackend) = 2π
 folding_limits(::FINUFFTBackend) = (-3π, 3π)  # we could even reduce this...
 non_uniform_type(::Type{T}, ::FINUFFTBackend) where {T <: AbstractFloat} = Complex{T}
 
+function init_fourier_vector_field(::FINUFFTBackend, ::Type{T}, Nks::Dims{M}) where {T <: Real, M}
+    # Data needs to be in a contiguous array of dimensions (Nks..., M) [usually M = 3].
+    data = Array{Complex{T}}(undef, Nks..., M)
+    components = ntuple(i -> view(data, :, :, :, i), Val(M))
+    uhat = StructArray{Vec3{Complex{T}}}(components)
+    @assert uhat isa StructArray{Vec3{Complex{T}}, 3}
+    @assert uhat.:1 isa SubArray
+    @assert parent(uhat.:1) === data  # we use this elsewhere to get the raw data before transforming
+    uhat
+end
+
 struct FINUFFTCache{
         T,
         CacheCommon <: LongRangeCacheCommon{T},
@@ -87,12 +98,13 @@ struct FINUFFTCache{
     common :: CacheCommon
     plan_type1 :: Plan  # plan for type-1 NUFFT (physical non-uniform → Fourier uniform)
     plan_type2 :: Plan  # plan for type-2 NUFFT (Fourier uniform → physical non-uniform)
+    charge_data :: Vector{Complex{T}}  # used to store charge data used in transforms
 end
 
 function init_cache_long_ewald(
         pc::ParamsCommon{T},
         params::ParamsLongRange{T, <:FINUFFTBackend}, args...,
-    ) where {T}
+    ) where {T <: AbstractFloat}
     (; Ls,) = pc
     (; backend, Ns,) = params
     n_modes = collect(Int64, Ns)  # type expected by finufft_makeplan
@@ -102,7 +114,8 @@ function init_cache_long_ewald(
     @assert Ns == Nks
     plan_type1 = _make_finufft_plan_type1(backend, n_modes, T)
     plan_type2 = _make_finufft_plan_type2(backend, n_modes, T)
-    FINUFFTCache(cache_common, plan_type1, plan_type2)
+    charge_data = Complex{T}[]
+    FINUFFTCache(cache_common, plan_type1, plan_type2, charge_data)
 end
 
 # FINUFFT options which should never be modified!
@@ -114,7 +127,7 @@ function _make_finufft_plan_type1(p::FINUFFTBackend, n_modes::Vector{Int64}, ::T
     opts = _finufft_options()
     type = 1
     iflag = -1  # this sets the FFT convention (we use a - sign for the forward transform)
-    ntrans = 1  # number of transforms to compute simultaneously (1 vector component at a time)
+    ntrans = 3  # number of transforms to compute simultaneously (3 vector components at once)
     finufft_makeplan(type, n_modes, iflag, ntrans, p.tol; dtype = T, p.kws..., opts...)
 end
 
@@ -122,39 +135,91 @@ function _make_finufft_plan_type2(p::FINUFFTBackend, n_modes::Vector{Int64}, ::T
     opts = _finufft_options()
     type = 2
     iflag = +1  # this sets the FFT convention (we use a + sign for the backward transform)
-    ntrans = 1  # number of transforms to compute simultaneously (1 vector component at a time)
+    ntrans = 3  # number of transforms to compute simultaneously (3 vector components at once)
     finufft_makeplan(type, n_modes, iflag, ntrans, p.tol; dtype = T, p.kws..., opts...)
 end
 
+function finufft_copy_charges_to_matrix!(
+        A::AbstractMatrix{<:Complex},
+        qs_in::StructVector{<:Vec3},
+    )
+    @assert axes(A, 1) == eachindex(qs_in)
+    @assert size(A, 2) == 3
+    qs = StructArrays.components(qs_in) :: NTuple{3}
+    @inbounds for (j, qj) ∈ pairs(qs)
+        for (i, q) ∈ pairs(qj)
+            A[i, j] = q
+        end
+    end
+    A
+end
+
+function finufft_copy_charges_from_matrix!(
+        qs_in::StructVector{<:Vec3},
+        A::AbstractMatrix{<:Complex},
+    )
+    @assert axes(A, 1) == eachindex(qs_in)
+    @assert size(A, 2) == 3
+    qs = StructArrays.components(qs_in) :: NTuple{3}
+    @inbounds for (j, qj) ∈ pairs(qs)
+        for i ∈ eachindex(qj)
+            # Note: qj may be a vector of real values, while A is a matrix of complex values.
+            # In our case we expect the complex part to be zero.
+            qj[i] = A[i, j]
+        end
+    end
+    qs_in
+end
+
+# Equivalent to reshape(v, :, M), but avoids issues when trying to resize `vs` later.
+function unsafe_reshape_vector_to_matrix(v::Vector, N, ::Val{M}) where {M}
+    @assert length(v) == N * M
+    p = pointer(v)
+    dims = (N, M)
+    unsafe_wrap(Array, p, dims; own = false)
+end
+
 function transform_to_fourier!(c::FINUFFTCache)
-    (; plan_type1,) = c
+    (; plan_type1, charge_data,) = c
     (; pointdata, uhat,) = c.common
     (; points, charges,) = pointdata
     # Interpret StructArrays as tuples of arrays (which is their actual layout).
-    points = StructArrays.components(points) :: NTuple{3, <:AbstractVector}
-    charges = StructArrays.components(charges)
-    uhat = StructArrays.components(uhat)
-    finufft_setpts!(plan_type1, points...)
-    for (qs, us) ∈ zip(charges, uhat)
-        @assert eltype(qs) <: Complex
-        finufft_exec!(plan_type1, qs, us)
-        _ensure_hermitian_symmetry!(c.common.wavenumbers, us)
+    points_data = StructArrays.components(points) :: NTuple{3, <:AbstractVector}
+    Np = length(charges)
+    @assert Np == length(points)
+    uhat_data = parent(uhat.:1)
+    T = eltype(uhat_data)
+    @assert T <: Complex
+    finufft_setpts!(plan_type1, points_data...)
+    resize!(charge_data, 3 * Np)
+    # Note: FINUFFT requires A to be an Array. This means that we can't use Bumper
+    # allocators here, as they return some other type of AbstractArray.
+    GC.@preserve charge_data begin
+        A = unsafe_reshape_vector_to_matrix(charge_data, Np, Val(3))
+        finufft_copy_charges_to_matrix!(A, charges)
+        finufft_exec!(plan_type1, A, uhat_data)  # execute NUFFT on all components at once
     end
+    _ensure_hermitian_symmetry!(c.common.wavenumbers, uhat)
     c
 end
 
 function interpolate_to_physical!(c::FINUFFTCache)
-    (; plan_type2,) = c
+    (; plan_type2, charge_data,) = c
     (; pointdata, uhat,) = c.common
     (; points, charges,) = pointdata
     # Interpret StructArrays as tuples of arrays (which is their actual layout).
-    points = StructArrays.components(points) :: NTuple{3, <:AbstractVector}
-    charges = StructArrays.components(charges)
-    uhat = StructArrays.components(uhat)
-    finufft_setpts!(plan_type2, points...)
-    for (qs, us) ∈ zip(charges, uhat)
-        @assert eltype(qs) <: Complex
-        finufft_exec!(plan_type2, us, qs)
+    points_data = StructArrays.components(points) :: NTuple{3, <:AbstractVector}
+    Np = length(charges)
+    @assert Np == length(points)
+    uhat_data = parent(uhat.:1)
+    T = eltype(uhat_data)
+    @assert T <: Complex
+    finufft_setpts!(plan_type2, points_data...)
+    resize!(charge_data, 3 * Np)
+    GC.@preserve charge_data begin
+        A = unsafe_reshape_vector_to_matrix(charge_data, Np, Val(3))
+        finufft_exec!(plan_type2, uhat_data, A)  # result is computed onto A
+        finufft_copy_charges_from_matrix!(charges, A)
     end
     c
 end